Removed deprecated docstring style, updating module naming convention, and changes tabs to spaces
